Assignment8/backend/app/api/routes.py
from fastapi import APIRouter, HTTPException, Depends
from typing import Dict, List
from ..services.model_service import ModelService
import logging
from pydantic import BaseModel

logger = logging.getLogger(__name__)
router = APIRouter()

# Initialize model service
model_service = ModelService()

class PredictionRequest(BaseModel):
    text: str

class PredictionResponse(BaseModel):
    prediction: int
    confidence: float
    inference_time: float
    timestamp: str

@router.post("/predict", response_model=PredictionResponse)
async def predict(request: PredictionRequest):
    """
    Make a prediction using the loaded model
    """
    try:
        result = await model_service.predict(request.text)
        return result
    except Exception as e:
        logger.error(f"Prediction error: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

@router.get("/metrics")
async def get_metrics():
    """
    Get model performance metrics
    """
    try:
        return model_service.get_metrics()
    except Exception as e:
        logger.error(f"Error fetching metrics: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

@router.get("/model/health")
async def model_health():
    """
    Check model health status
    """
    try:
        return model_service.health_check()
    except Exception as e:
        logger.error(f"Health check error: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

# Training endpoints (for demonstration)
@router.post("/model/train")
async def train_model():
    """
    Simulate model training (placeholder)
    """
    return {
        "status": "success",
        "message": "Model training initiated",
        "details": {
            "training_job_id": "demo_job_001",
            "status": "running",
            "progress": "0%"
        }
    }

@router.get("/model/training-status/{job_id}")
async def get_training_status(job_id: str):
    """
    Get training job status (placeholder)
    """
    return {
        "job_id": job_id,
        "status": "completed",
        "progress": "100%",
        "metrics": {
            "accuracy": 0.92,
            "loss": 0.08,
            "epochs_completed": 10
        }
    }

# Data pipeline endpoints (for demonstration)
@router.get("/data/status")
async def get_data_pipeline_status():
    """
    Get data pipeline status (placeholder)
    """
    return {
        "status": "active",
        "last_update": "2023-11-15T10:00:00Z",
        "processed_records": 10000,
        "data_quality_score": 0.95,
        "latest_batch": {
            "id": "batch_001",
            "size": 1000,
            "status": "processed"
        }
    }
