Assignment8/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
datasets==2.14.6
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
pytest==7.4.3
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psutil==5.9.6
prometheus-client==0.19.0
